services/ml-inference/requirements.txt
flask==3.0.3
gunicorn==22.0.0
kafka-python==2.0.2
prometheus-client==0.20.0
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
